shared: Output alias if the type already exists
Makes it so that even if there would otherwise be no code to generate,
if you say you want a type Foobar, you can still refer to Foobar,
instead of getting an error.

The name will still coerced to a valid type name though. Something to
consider for the rewrite, if that is a good idea at all.

Closes #18

Co-authored-by: Gygaxis Vainhardt <gygaxis@zingolabs.com>