Use requested name for root type or type alias
Fixes #20

BREAKING: Requesting the name Foo for a list of objects would previously
use the name Foo for the type of the elements. It will now use Foo for
the list (the root type). If the codegeneration is able to "singularize"
the name (e.g. "Points" to "Point") the object type name will be the
same as before ("Point") and the plural will be used for the list.