shared: Remove more dependencies. Inflector, regex, lazy_static
regex: Due to stuff like unicode tables this actually adds a lot of
bloat to the wasm bundle. After this commit wasm bundle size is ~300 KB.
Regexes were only used in a few easy to replace locations.
Fixes #14.

Inflector: Was used for singularizing terms for type names of array
elements. Uses regex, so had to be removed for removing regex to make
sense. Was never all that happy with the results anyway. Currently
replaced with a placeholder implementation. Plan is to expand that, but
conservatively. I.e. not singularizing a term is better than incorrectly
trying to singularize something.

lazy_static: For the non-regex uses (lists of reserved words) it seemed
to not be needed anymore, as I could just do `.contains()` directly on
`&[&str]`. Maybe that was always possible and it was just prematurely
optimized, but removed it now at least.