Detect and reuse duplicate named types.
Fixes #3

This is far from perfect, but it at least handles the simple cases. It
requires shapes to be exactly equal, including field order. This can be
fixed relatively easily, but some cases will require a significant
refactor (which I've started / made proof-of-concept-versions of in the
past, but not finished). In particular if shapes are slightly different,
and the already created code is not usable for the one we are currently
looking at (e.g. using integers, but new type requires floats).