shared: Remove option runnable
Isn't really needed for any usecase, and was only possible to use by
using json_typegen_shared as a library, and expliitly enabling the
`runnable` option. Since shared as a library used to be discouraged,
this should hopefully not be relevant to anyone.

BREAKING: Option `runnable` has been completely removed.